+++ packages/patternfly-3/patternfly-react/src/index.js
@@ -17,6 +17,7 @@ export * from './components/Filter';
 export * from './components/Form';
 export * from './components/Grid';
 export * from './components/HintBlock';
+export * from './components/HorizontalNav';
 export * from './components/Icon';
 export * from './components/InfoTip';
 export * from './components/InlineEdit';