+++ packages/patternfly-4/react-core/src/components/Form/Form.d.ts
@@ -1,9 +1,0 @@
-import { FunctionComponent, HTMLProps } from 'react';
-
-export interface FormProps extends HTMLProps<HTMLFormElement> {
-    isHorizontal?: boolean;
-}
-
-declare const Form: FunctionComponent<FormProps>;
-
-export default Form;