+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.8...@patternfly/react-docs@4.4.9) (2019-05-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.4.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.7...@patternfly/react-docs@4.4.8) (2019-05-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.4.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.6...@patternfly/react-docs@4.4.7) (2019-05-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.4.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.5...@patternfly/react-docs@4.4.6) (2019-05-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.4.5](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.4...@patternfly/react-docs@4.4.5) (2019-05-17)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.4.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.3...@patternfly/react-docs@4.4.4) (2019-05-17)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.4.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.2...@patternfly/react-docs@4.4.3) (2019-05-17)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.4.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.4.1...@patternfly/react-docs@4.4.2) (2019-05-17)
 
 **Note:** Version bump only for package @patternfly/react-docs